backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pandas
openpyxl
python-multipart
passlib[bcrypt]
python-jose
python-dotenv
psycopg2-binary
